[ADD] product_main_on_order|sale: init
Adds a model that allows to create rules to identify which of the
multiple products on related order (can be any type of order) is main
one.

Sample use case: you want to promote specific product from order, like
including its description in printed document, but do not want to hard
code it.

NOTE. This is technical module meant to be used in tandem with some more
specific modules that can use this functionality.