[IMP] delivery_3pl: force warehouse
Possibility to force warehouse even if it was specified in vals. This
can be useful when default warehouse is set directly in vals and it
can't be distinguished from actually specified manually. Though of
course this can make it harder to specify warehouse you want.

Also added mechanism to allow registering specific 3PL services, to be
able to recognize which one needs to be used.