[FIX] delivery_3pl: force_picking_done
Redesigned force picking to handle both moves/lines that are already
reserved and those that need to be fully forced. Before it could handle
only one or the other.

Also, now when forcing picking to be done, we make sure to send email
about shipment, only if picking is really in 'done' state, to not send
it when its not supposed to.