Stamp configurator imp2 ala (#29)
* [ADD] product_stamp_configurator: stamp_type

Product template stamp_type use nearest category stamp type instead of
direct category only. This allows for better flexibility when there are
multiple levels of categories and stamp type is only set on some parent
category.

* [ADD] product_stamp_configurator: default mold qty

Most of the time mold quantity is the same.

* [ADD] product_stamp_configurator: die_default_id

Making possible to specify default die type on configurator as most of
the time it will be same type.

* [IMP] product_stamp_configurator: invisible categs

Hiding mold and counter die category fields on configurator if those
categories are already set (e.g. using defaults) as these are mostly
technical fields not really relevant for user.

* [FIX] product_stamp_configurator: omit False ref

When there is no ref, we should not blindly include it in code as it
would translate to "False" string.

* [IMP] product_stamp_configurator_sale: ref qty

Now we can get suggested dies quantity from selected insert die product
sale order line. It assumes that there is only single sale order line
for that product.

* [ADD] product_category_hs_code: init

Adds HS Code field on product category. When creating product, it will
use its category ``HS Code`` value if not specified on product.

* [ADD] product_stamp_configurator: msg cfg params

Now posting message on product/product template that lists parameter
values used to create those products using configurator.

* [ADD] product_stamp_configurator: min_area

If entered area in configurator is very low, we can now use minimum
area from pricelist, to force for price calculation, because cost is
not reduced when stamps are very small.

* [IMP] product_stamp_configurator: DP Stamp Weight

We can now use decimal precision for stamp weight_coefficient fields.

* [IMP] product_stamp_configurator: mold free

Now when `quantity_die_mold_free` is set to zero, it will be disabled,
so it would be more flexible when some pricelists don't need to use
this.