Stamp configurator imp3 ala (#30)
* [IMP] product_stamp_configurator: onchanges

Added onchanges to clean up data that is hidden when other field are
changed.

* [IMP] product_stamp_configurator: embossed

Previously used `is_embossed` field name was not entirely correct, so we
renamed it to `flat_embossed_foiling`. And now we added new field which
is called `is_embossed`, that is used to control whether we need
mold/counter dies.