Stamp configurator imp4 ala (#31)
[IMP] product_stamp_configurator: prices, naming

Redesigned the way price was calculated for dies. Initial used formulas
were incorrect as it was using stamp pricelist only if
flat_embossed_foiling was used, when it actually had to use in all
cases. Also material price was used twice in formula, when it had to
be used only once. Formulas are also simplified now.

Also changed some labels to make it easier to understand what it means
and now all prices use `Product Price` decimal.precision to be
configurable.

Now when we change quantity_dies on configurator, it also sets quantity
on quantity_counter_dies to match it.

Also, added more appropriate error when category is not specified on
configurator, instead of just throwing ensure_one error.