[FIX] product_stamp_configurator: _calc_weight
design object was being passed instead of expected material object, thus
weight was calculated incorrectly for both die and mold products.

We also now renamed `weight_coefficient` field to `weight` on
`material.stamp`, because it is weight of kg/cm², not a coefficient.

Also adding `recursive=True` on nearest_stamp_type field, because we
are doing recursive computation (to avoid warnings).