[REF] api_base: get_component_service, x2m
Improved x2m command, to be able to not iterate over source value when
its not needed (or when src value is not even iterable).

Also added utility function to make it easier to get component from any
collection.

get_orm_map is now expected to have list of tuples instead of
dictionary, so it would be possible to map same source key with multiple
destination fields.