[FIX] procurement_group_parent_root: self group
Apparently self related move can expose procurement group to be itself,
so it would end up with infinite loop. To make sure that won't happen,
we explicitly exclude it when looking for parent procurement group.

Now after having this fixed, it should be possible to compute
root procurement group on install as well.