[FIX] package_configurator: temp multi comp
When box component data is changed and not saved, it can "create"
multiple one in memory and we expect only one. So temp fix is to just
use first found when using layout.

[IMP] package_configurator: print only wrappingpaper

Normally only wrappingpaper is to be printed, so making that field
readonly for other scopes. It should be improved with proper constraint
as it would be possible to "hack" it and still set color on greyboard
components.