[ADD] stock_chain_auto_done: init
Auto finish Chained pickings by a rule.

When one picking is done and there is another in a chain, that matches
all chained moves, it will attempt to auto finish the next picking, but
only if all goods are available.

Chain rule can specify location to match in both source and
destination between chained moves. If all chained moves match this
location, action is triggered to finish the next picking.